Update CSharp/Nancy
- At least nancy-mono runs now.
- Added plaintext route test.
- Separate query test from db test.
- Temporarily remove libevent2 test. (Currently, this one does not work.)
