+++ frameworks/CSharp/nancy/NancyModules/NancyModules.csproj
@@ -52,8 +52,10 @@
     <Reference Include="System.Data" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="PlainModule.cs" />
     <Compile Include="DbModule.cs" />
     <Compile Include="JsonModule.cs" />
+    <Compile Include="QueryModule.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
