+++ frameworks/CSharp/nancy/NancyModules/PlainModule.cs
@@ -1,0 +1,15 @@
+namespace NancyModules 
+{
+  using Nancy;
+
+  public class PlainModule : NancyModule
+  {
+    public PlainModule() : base("plaintext")
+    {
+      Get["/"] = x =>
+      {
+        return "Hello, World!";
+      };
+    }
+  }
+}
