+++ frameworks/CSharp/nancy/benchmark_config.json
@@ -3,9 +3,10 @@
   "tests": [{
     "default": {
       "setup_file": "setup_iis",
+      "plaintext_url":"/plaintext",
       "json_url": "/json",
       "db_url": "/db",
-      "query_url": "/db/",
+      "query_url": "/query/",
       "port": 8080,
       "approach": "Realistic",
       "classification": "Micro",
@@ -23,9 +24,10 @@
     },
     "mono": {
       "setup_file": "setup_nginx",
+      "plaintext_url":"/plaintext",
       "json_url": "/json",
       "db_url": "/db",
-      "query_url": "/db/",
+      "query_url": "/query/",
       "port": 8080,
       "approach": "Realistic",
       "classification": "Micro",
@@ -40,26 +42,7 @@
       "display_name": "nancy",
       "notes": "",
       "versus": ""
-    },
-    "libevent2": {
-      "setup_file": "setup_libevent",
-      "json_url": "/json",
-      "db_url": "/db",
-      "query_url": "/db/",
-      "port": 8080,
-      "approach": "Realistic",
-      "classification": "Micro",
-      "database": "MySQL",
-      "framework": "nancy",
-      "language": "C#",
-      "orm": "Raw",
-      "platform": "Mono",
-      "webserver": "nginx",
-      "os": "Linux",
-      "database_os": "Linux",
-      "display_name": "nancy-libevent",
-      "notes": "",
-      "versus": ""
     }
+
   }]
 }
