+++ frameworks/CSharp/nancy/nginx.conf
@@ -19,7 +19,7 @@ http {
 
         location / {
             fastcgi_pass mono;
-            include /usr/local/nginx/conf/fastcgi_params;
+            include nginx.osenv.conf; # read os env from this file
             fastcgi_param  SCRIPT_FILENAME $document_root$fastcgi_script_name;
         }
     }
