+++ frameworks/CSharp/nancy/setup_nginx.sh
@@ -19,6 +19,13 @@ for port in $(seq ${port_start} $port_end); do
 done
 conf+="}"
 
+# Store path of fastcgi_params dynamically in a file called "nginx.osenv.conf". 
+# The reason why I do this is Ngix "include" cannot recognize variables. (Well known issue of Nginx)
+# To use OS Environment at ngix.conf 3rd party library or perl module is needed
+# Current approach is one trick to solve the problem without utilize those 3rd party libraries.
+echo "include $IROOT/nginx/conf/fastcgi_params;" > $TROOT/nginx.osenv.conf
+
+
 echo -e $conf > $TROOT/nginx.upstream.conf
 ${NGINX_HOME}/sbin/nginx -c $TROOT/nginx.conf -g "worker_processes '"${MAX_THREADS}"';"
